vadstena/vts: WIP
git-svn-id: https://scm.citationtech.net/trunk/vadstena@14156 b7e6147c-0f76-4c84-a09d-19069afaf10c

+++ vts-libs/storage/lod.hpp
@@ -5,7 +5,8 @@
 
 namespace vadstena { namespace storage {
 
-typedef std::uint8_t Lod;
+// not using std::uint8_t because it is interpreted as char
+typedef std::uint16_t Lod;
 
 } } // namespace vadstena::storage
